+++ lib/src/simplified_builder.dart
@@ -2,3 +2,6 @@ library simplified_builder;
 
 export 'simplified_future_builder.dart';
 export 'simplified_stream_builder.dart';
+export 'simplified_action_menu_builder.dart';
+export 'simplified_grid_data_builder.dart';
+export 'simplified_form_builder.dart';