+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.2
 
 * Simplifies the usage of FutureBuilder and StreamBuilder.
+* SimplifiedActionMenuBuilder is addded for simplified action menu creation.
+* SimplifiedGridDataBuilder is addded for simplified grid data presentation.
+* SimplifiedFormBuilder for is addded simplified form creation.
+